rename finishCondition to match default export naming
