+++ frontend/src/BrowseInventory/BrowseInventoryRow.tsx
@@ -5,7 +5,7 @@ import Price from '../common/Price';
 import CardImageTooltip from '../ui/CardImageTooltip';
 import SetIcon from '../ui/SetIcon';
 import { Finish } from '../utils/ClientCard';
-import displayFinishCondition from '../utils/finishCondition';
+import displayFinishCondition from '../utils/displayFinishCondition';
 import { ResponseCard } from './filteredCardsQuery';
 
 interface Props {
