+++ frontend/src/common/QohLabels.tsx
@@ -2,7 +2,7 @@ import { Tooltip, Typography } from '@material-ui/core';
 import React, { FC } from 'react';
 import InventoryChip from '../ui/InventoryChip';
 import { QOH } from '../utils/ClientCard';
-import displayFinishCondition from '../utils/finishCondition';
+import displayFinishCondition from '../utils/displayFinishCondition';
 import parseQoh from '../utils/parseQoh';
 
 export const createInventoryLineItems = (
