+++ frontend/src/utils/displayFinishCondition.ts
